Create e2e/test_home.py
